uses suspend and start to satisfy race detector, check output to ensure wait is working
